+++ src/main/java/com/itwillbs/domain/MaterialOutboundItemVO.java
@@ -1,5 +1,7 @@
 package com.itwillbs.domain;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -8,7 +10,7 @@ public class MaterialOutboundItemVO {
     private String outboundItemId;    // 출고 상세 ID (출고 항목 고유 식별자)
     private String outboundId;        // 출고 ID (연결된 출고 ID)
     private String materialId;        // 자재 ID (출고된 자재 ID)
-    private int quantity;             // 출고 수량
+    private BigDecimal quantity;      // 출고 수량
     private String lotNo;             // LOT 번호 (출고 자재의 LOT 번호 - 선택사항)
     private String inventoryId;
     private int requiredQty;          // 필요 수량 (출고 요청 수량)
